drm: xilinx: Fix VDMA compatibility
Commit 3d6d7111c6a90f2e0dbc8a7da4ad7e1cd65399b0 broke
compatibility of the Xilinx DRM driver with VDMA IPs that require
the format be specified in the device tree. This fix handles the
case when it is identified that the DMA is not the new framebuffer
IP and allows the format to be set from the device tree.

RPH: This update is from Sam Bobrowicz