Added driver to interface with 7-Segment display for Xilinx FPGAs.
The 7-Segment driver is designed to interface with the Real Digital
display_7seg_cntr_1.0 hardware IP for Xilinx FPGAs.  The hardware
also includes up to two frequency counter inputs that can be used.
The registers in the hardware are accessible as a UIO device in user
space, which allows for easy access.